+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Base class for excluded parameters.
 - Resolving of asset/hierarchy id within tools.
 
+### Fixed
+- frontend link in ion channel recording tool.
+
 ## [v0.11.2] - 3.11.2025
 
 ### Fixed