+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.11.3] - 3.11.2025
+
 ### Fixed
 - frontend link in ion channel recording tool.
+- Allow null inputs in connectivity tool.
 
 ## [v0.11.2] - 3.11.2025